+++ scout/utils/broad_liftover_client.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Optional
+from typing import Optional, Tuple
 
 import requests
 from flask import flash